requirements.txt
beautifulsoup4
confidence
facebook-scraper
google-api-python-client
lxml_html_clean
nltk
numpy
openpyxl
scipy
striprtf
spacy
pandas
pyyaml
requests
tqdm